Restructured the project directory
